feat(extensions): notes is the worked example of all five methods
Every extension operation was a POST because the seam admitted nothing else.
Now the method follows the operation's own semantic, and notes — the unit other
authors read — shows each case with the reasoning attached:

  GET    /ext/notes/list                 zero-argument read
  GET    /ext/notes/signing-key/status   zero-argument read
  POST   /ext/notes/add                  create, arguments in the body
  PUT    /ext/notes/signing-key          replace: the same call twice, same state
  DELETE /ext/notes/remove?id=…          the query-argument case, end to end
  POST   /ext/notes/signature            a READ that stays a POST, deliberately

The last row is the one worth reading. A GET would put the signed payload in the
URL, and a URL reaches the access log, browser history and every proxy between
— this repo already redacts publicPreferencesPrefix for exactly that reason. The
4096-character limit settles it independently, since a request line is capped far
lower than a body. The cost is named in the fragment: because it stays a POST the
seat ceiling classifies it as mutating, so a read seat cannot call a read. Fixing
THAT needs classification from the declared scope, which these method rules only
narrow.

PATCH stays unused: notes has no edit-a-note operation, and inventing one to
complete the set would be an operation existing for a demo rather than a user.

yogi and crm-hello become GETs on the same grounds — both are zero-argument
reads. Unit manifests and the composition regenerate; the digests move because
the method is part of the published operation's identity, which is what stops an
operator resolution recorded against one method carrying silently to another.

Also here, both found by the gates rather than by taste:

  - extverbs.go passed the 500-LOC cap, so the schema-reading half moved to
    extverbschemas.go. The two answer different questions — an operation's
    identity and governance, versus where its arguments live and how a declared
    schema becomes a standalone literal.
  - the argument source is now a function resolved at mount (argumentReader)
    rather than a nil-means-body sentinel the handler branched on. A route holds
    only the reader its own declaration produced, so it cannot read a body it has
    no schema for. The handler loses the branch entirely.

Refs #993
