+++ extensions/yogi/manifest.generated.json
@@ -11,13 +11,13 @@
       "operation": "agent.tool.invoke",
       "operation_id": "yogiQuote",
       "route": "/ext/yogi/quote",
-      "method": "POST",
+      "method": "GET",
       "scopes": [
         "read"
       ],
       "tier": "auto_execute",
-      "fragment_hash": "0683ac243f03fcd05521ba9c5f5fd0c5c62557e23fd65021821cfa9bef18f1f1",
-      "digest": "sha256:5734dc8ed99fc6c2178fdacc3a532000e81dddb075b21f0a8dff7ba78b0e7598"
+      "fragment_hash": "00c2cd8d6ef6bb1c962b6016ee6cd90c22e55688256ebfa108cc917d5bd138e9",
+      "digest": "sha256:ee6acae9433b8fa3173f2aa6c6971dbde2954c2f0ee92b90a981d4edf9d1da6d"
     }
   ]
 }
